Add write(ASInputStream) to COSFilterFlateEncode

+++ src/main/java/org/verapdf/cos/filters/COSFilterFlateDecode.java
@@ -68,7 +68,7 @@ public class COSFilterFlateDecode extends ASBufferedInFilter {
             this.bufferSize = bytesFed;
             inflater.setInput(this.buffer, 0, this.bufferSize);
         }
-        long startOffset = this.bufferSize - inflater.getRemaining();
+        int startOffset = this.bufferSize - inflater.getRemaining();
         try {
             int res = inflater.inflate(buffer, 0, size);
             if (res == 0) {
@@ -85,7 +85,7 @@ public class COSFilterFlateDecode extends ASBufferedInFilter {
             }
         } catch (DataFormatException e) {
             try {
-                return readByByte(buffer, (int)startOffset, size);
+                return readByByte(buffer, startOffset, size);
             } catch (IOException exp) {
                 throw new IOException("Can't decode Flate encoded data", e);
             }
